cleanup and reformat completed for release

+++ cvsu_memory.h
@@ -8,20 +8,20 @@
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *   * Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *   * Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in the
- *     documentation and/or other materials provided with the distribution.
- *   * Neither the name of the copyright holder nor the
- *     names of its contributors may be used to endorse or promote products
- *     derived from this software without specific prior written permission.
+ *   * Redistributions of source code must retain the above copyright notice,
+ *     this list of conditions and the following disclaimer.
+ *   * Redistributions in binary form must reproduce the above copyright notice,
+ *     this list of conditions and the following disclaimer in the documentation
+ *     and/or other materials provided with the distribution.
+ *   * Neither the name of the copyright holder nor the names of its
+ *     contributors may be used to endorse or promote products derived from this
+ *     software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL COPYRIGHT HOLDER BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * ARE DISCLAIMED. IN NO EVENT SHALL COPYRIGHT HOLDER BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
  * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
  * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
  * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
@@ -41,36 +41,40 @@ extern "C" {
 /**
  * A generic function for allocating an array of bytes.
  */
-result memory_allocate(
-    data_pointer *target,
-    size_t target_size,
-    size_t element_size
+result memory_allocate
+(
+  data_pointer *target,
+  uint32 target_size,
+  uint32 element_size
 );
 
 /**
  * A generic function for deallocating an array of bytes.
  */
-result memory_deallocate(
-    data_pointer *target
+result memory_deallocate
+(
+  data_pointer *target
 );
 
 /**
  * A generic function for resetting an array of bytes.
  */
-result memory_clear(
-    data_pointer target,
-    size_t target_size,
-    size_t element_size
+result memory_clear
+(
+  data_pointer target,
+  uint32 target_size,
+  uint32 element_size
 );
 
 /**
  * A generic function for copying an array of bytes.
  */
-result memory_copy(
-    data_pointer target,
-    const data_pointer source,
-    size_t copy_size,
-    size_t element_size
+result memory_copy
+(
+  data_pointer target,
+  const data_pointer source,
+  uint32 copy_size,
+  uint32 element_size
 );
 
 #ifdef __cplusplus
